调试数据查询部分

+++ management-dao/src/main/java/com/fooee/winqing/management/dao/vdo/book/BookDescriptionInfoDo.java
@@ -10,12 +10,12 @@ public class BookDescriptionInfoDo extends QueryResult{
     /**
      *
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 图书ID
      */
-    private Integer bookId;
+    private Long bookId;
 
     /**
      * 目录信息
@@ -50,19 +50,19 @@ public class BookDescriptionInfoDo extends QueryResult{
         this.mediaCommentInfo = mediaCommentInfo;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getBookId() {
+    public Long getBookId() {
         return bookId;
     }
 
-    public void setBookId(Integer bookId) {
+    public void setBookId(Long bookId) {
         this.bookId = bookId;
     }
