+++ management-dao/src/main/java/com/fooee/winqing/management/dao/vdo/passport/ManageDepartmentFunctionRelationDo.java
@@ -10,15 +10,15 @@ public class ManageDepartmentFunctionRelationDo {
 	/**
 	 * 组织机构id
 	 */
-	private Integer departmentId;
+	private Long departmentId;
 	/**
 	 * 功能id
 	 */
-	private Integer functionId;
+	private Long functionId;
 	/**
 	 * 
 	 */
-	private Integer id;
+	private Long id;
 	/**
 	 * 是否启用
 	 */
@@ -28,14 +28,14 @@ public class ManageDepartmentFunctionRelationDo {
 	 * @param departmentId
 	 */
 	
-	public void setDepartmentId(Integer departmentId){
+	public void setDepartmentId(Long departmentId){
 		this.departmentId=departmentId;
 	}
 	/**
 	 * 组织机构id
 	 * @return
 	 */
-	public Integer getDepartmentId(){
+	public Long getDepartmentId(){
 		return this.departmentId;
 	}
 		
@@ -44,14 +44,14 @@ public class ManageDepartmentFunctionRelationDo {
 	 * @param functionId
 	 */
 	
-	public void setFunctionId(Integer functionId){
+	public void setFunctionId(Long functionId){
 		this.functionId=functionId;
 	}
 	/**
 	 * 功能id
 	 * @return
 	 */
-	public Integer getFunctionId(){
+	public Long getFunctionId(){
 		return this.functionId;
 	}
 		
@@ -60,14 +60,14 @@ public class ManageDepartmentFunctionRelationDo {
 	 * @param id
 	 */
 	
-	public void setId(Integer id){
+	public void setId(Long id){
 		this.id=id;
 	}
 	/**
 	 * 
 	 * @return
 	 */
-	public Integer getId(){
+	public Long getId(){
 		return this.id;
 	}
