+++ management-dao/src/main/java/com/fooee/winqing/management/dao/vdo/passport/ManageFunctionDo.java
@@ -24,7 +24,7 @@ public class ManageFunctionDo {
 	/**
 	 *
 	 */
-	private Integer id;
+	private Long id;
 	/**
 	 * 是否启用
 	 */
@@ -32,7 +32,7 @@ public class ManageFunctionDo {
 	/**
 	 * 父级id
 	 */
-	private Integer parentId;
+	private Long parentId;
 
 	/**
 	 * 树形子集
@@ -101,14 +101,14 @@ public class ManageFunctionDo {
 	 * @param id
 	 */
 
-	public void setId(Integer id){
+	public void setId(Long id){
 		this.id=id;
 	}
 	/**
 	 *
 	 * @return
 	 */
-	public Integer getId(){
+	public Long getId(){
 		return this.id;
 	}
 
@@ -133,14 +133,14 @@ public class ManageFunctionDo {
 	 * @param parentId
 	 */
 
-	public void setParentId(Integer parentId){
+	public void setParentId(Long parentId){
 		this.parentId=parentId;
 	}
 	/**
 	 * 父级id
 	 * @return
 	 */
-	public Integer getParentId(){
+	public Long getParentId(){
 		return this.parentId;
 	}
