+++ management-dao/src/main/java/com/fooee/winqing/management/dao/vdo/passport/ManageFunctionQc.java
@@ -22,7 +22,7 @@ public class ManageFunctionQc{
 	/**
 	 * 
 	 */
-	private Integer id;
+	private Long id;
 	/**
 	 * 是否启用
 	 */
@@ -30,7 +30,7 @@ public class ManageFunctionQc{
 	/**
 	 * 父级id
 	 */
-	private Integer parentId;
+	private Long parentId;
 	/**
 	 * 功能描述
 	 * @param functionDescription
@@ -84,14 +84,14 @@ public class ManageFunctionQc{
 	 * @param id
 	 */
 	
-	public void setId(Integer id){
+	public void setId(Long id){
 		this.id=id;
 	}
 	/**
 	 * 
 	 * @return
 	 */
-	public Integer getId(){
+	public Long getId(){
 		return this.id;
 	}
 		
@@ -116,14 +116,14 @@ public class ManageFunctionQc{
 	 * @param parentId
 	 */
 	
-	public void setParentId(Integer parentId){
+	public void setParentId(Long parentId){
 		this.parentId=parentId;
 	}
 	/**
 	 * 父级id
 	 * @return
 	 */
-	public Integer getParentId(){
+	public Long getParentId(){
 		return this.parentId;
 	}
