+++ management-dao/src/main/java/com/fooee/winqing/management/dao/vdo/passport/ManageRoleQc.java
@@ -10,7 +10,7 @@ public class ManageRoleQc{
 	/**
 	 * 
 	 */
-	private Integer id;
+	private Long id;
 	/**
 	 * 是否启用
 	 */
@@ -28,14 +28,14 @@ public class ManageRoleQc{
 	 * @param id
 	 */
 	
-	public void setId(Integer id){
+	public void setId(Long id){
 		this.id=id;
 	}
 	/**
 	 * 
 	 * @return
 	 */
-	public Integer getId(){
+	public Long getId(){
 		return this.id;
 	}
